requirements.txt
# Gemini-Only Requirements for Agentic RAG Knowledge Graph
# Based on working graphiti-agent setup

# Core dependencies for Graphiti with Gemini
graphiti-core[google-genai]==0.18.0
google-genai>=1.16.1

# Core Python dependencies
python-dotenv>=1.1.0
pydantic>=2.11.0
pydantic-ai>=0.2.9
rich>=14.0.0

# Database and graph dependencies
neo4j>=5.28.0
asyncpg>=0.30.0

# HTTP and async dependencies
httpx>=0.28.0
aiohttp>=3.12.0
tenacity>=9.0.0

# Utility dependencies
click>=8.2.1
typing-extensions>=4.12.0
diskcache>=5.6.3

# Data processing
numpy>=2.3.0
PyYAML>=6.0.2

# Testing (optional)
pytest>=8.4.0
pytest-asyncio>=1.0.0

# Development utilities (optional)
colorama>=0.4.6
tqdm>=4.67.0
